Attempting a static files howto rewrite.
The current howto has some issues ( https://code.djangoproject.com/ticket/19582 ) and doesn't clearly explain what staticfiles is doing and why. For those coming at staticfiles with the expectation of "put files in this folder, they get served at this URL", the app-specific static files idea is confusing. This rewrite makes it more clear what's necessary to get that outcome, and explain why the process is the way it is.